requirements.txt
# 超級安全版本 - 避免所有相容性問題
flask==2.2.5
requests==2.28.2
gunicorn==20.1.0
python-dotenv==0.21.1

# 基本工具
schedule==1.2.0

# Line Bot 功能
line-bot-sdk==2.4.2

# Firebase 功能
firebase-admin==5.4.0

# 爬蟲功能
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 日期處理
python-dateutil==2.8.2 